# [PowerToys Run] Windows Terminal plugin: Improve query speed