--- a/src/modules/launcher/Plugins/Microsoft.PowerToys.Run.Plugin.WindowsTerminal/Helpers/TerminalQuery.cs
+++ b/src/modules/launcher/Plugins/Microsoft.PowerToys.Run.Plugin.WindowsTerminal/Helpers/TerminalQuery.cs
@@ -10,6 +10,7 @@ using System.IO;
 using System.Linq;
 using System.Security.Principal;
 using System.Threading;
+using Windows.ApplicationModel;
 using Windows.Foundation;
 using Windows.Management.Deployment;
 
@@ -17,7 +18,8 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsTerminal.Helpers
 {
     public class TerminalQuery : ITerminalQuery
     {
-        private readonly PackageManager _packageManager;
+        // Cache for the package AUMID that we don't have to query the AUMID every time. (On slow systems, each query costs up to 50 ms.)
+        private static readonly Dictionary<string, AumidCacheData> _PackageAumidCache = new Dictionary<string, AumidCacheData>();
 
         private static ReadOnlyCollection<string> Packages => new List<string>
         {
@@ -25,8 +27,21 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsTerminal.Helpers
             "Microsoft.WindowsTerminalPreview",
         }.AsReadOnly();
 
-        // Cache for the package AUMID that we don't have to query the AUMID every time. (On slow systems, each query costs up to 50 ms.)
-        private static readonly Dictionary<string, string> _PackageAumidCache = new Dictionary<string, string>();
+        // Struct for value data in AUMID cache
+        private struct AumidCacheData
+        {
+            public PackageVersion CachedPackageVersion { get; private set; }
+
+            public string PackageAumid { get; private set; }
+
+            public AumidCacheData(PackageVersion v, string id)
+            {
+                CachedPackageVersion = v;
+                PackageAumid = id;
+            }
+        }
+
+        private readonly PackageManager _packageManager;
 
         private IEnumerable<TerminalPackage> Terminals => GetTerminals();
 
@@ -60,7 +75,7 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsTerminal.Helpers
 
             foreach (var p in _packageManager.FindPackagesForUser(user.Value).Where(p => Packages.Contains(p.Id.Name)))
             {
-                if (!_PackageAumidCache.ContainsKey(p.Id.Name))
+                if (!_PackageAumidCache.ContainsKey(p.Id.Name) || !_PackageAumidCache[p.Id.Name].CachedPackageVersion.Equals(p.Id.Version))
                 {
                     // When changing the target release to 19041 in the future, we can use <GetAppListEntries()> instead of <GetAppListEntriesAsync()>!
                     var appListEntries = p.GetAppListEntriesAsync();
@@ -70,14 +85,14 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsTerminal.Helpers
                         Debug.Print($"Sleep 10 sec for AUMID for {p.Id.Name}>");
                     }
 
-                    _PackageAumidCache.Add(p.Id.Name, appListEntries.GetResults().Single().AppUserModelId);
+                    _PackageAumidCache[p.Id.Name] = new AumidCacheData(p.Id.Version, appListEntries.GetResults().Single().AppUserModelId);
                 }
 
-                var aumid = _PackageAumidCache[p.Id.Name];
+                var aumid = _PackageAumidCache[p.Id.Name].PackageAumid;
                 var version = new Version(p.Id.Version.Major, p.Id.Version.Minor, p.Id.Version.Build, p.Id.Version.Revision);
                 var settingsPath = Path.Combine(localAppDataPath, "Packages", p.Id.FamilyName, "LocalState", "settings.json");
                 yield return new TerminalPackage(aumid, version, p.DisplayName, settingsPath, p.Logo.LocalPath);
             }
-        }
+      }
     }
 }