## Summary of the Pull Request

**What is this about:**
Before this change we query the AppUserModellID (AUMID) for every installed terminal app on every query. Because the method we used is async we sleeped 100 ms in a loop while waiting for the result. This leads to high query times: On each query up to 250 ms when wt and wt prev are installed.

I have changed the code and now we have query times up to 15 ms on fast machines and up to 35 ms on slow machines.

**What is included in the PR:** 
Now that we [know](https://github.com/microsoft/PowerToys/pull/15836#issuecomment-1025204301) that the AUMID didn't change in the future we can write it fixed in the code. Then we don't have to query it and the plugin queries become faster.

I have changed the package list into a dictionary and it now contains the packages names and the corresponding AUMIDs.

**How does someone test / validate:** 
Run local dev build to test the plugin function and checked the log for errors and query times.

## Quality Checklist

- [x] **Linked issue:** #15778
- [x] **Communication:** I've discussed this with core contributors in the issue. 
- [x] **Tests:** Added/updated and all pass
- [ ] **Installer:** Added/updated and all pass
- [ ] **Localization:** All end user facing strings can be localized
- [ ] **Docs:** Added/ updated
- [x] **Binaries:** Any new files are added to WXS / YML
   - [x] No new binaries
   - [ ] [YML for signing](https://github.com/microsoft/PowerToys/blob/main/.pipelines/pipeline.user.windows.yml#L68) for new binaries
   - [ ] [WXS for installer](https://github.com/microsoft/PowerToys/blob/main/installer/PowerToysSetup/Product.wxs) for new binaries

## Contributor License Agreement (CLA)
A CLA must be signed. If not, go over [here](https://cla.opensource.microsoft.com/microsoft/PowerToys) and sign the CLA.