@htcfreek 2022-01-29
> Tested and working as expected!
> Not sure if there are any side effects on caching AUMID.

As I know it doesn't change as long as the app exists in the store.

AUMID for wt is something like `Microsoft.WindowsTerminal_*` and for wt preview it's the same, but with `WindowsTerminalPreviw`.