@htcfreek 2022-01-29
@davidegiacometti 
The syntax of AMUID is `<package.Id.FamilyName>!App` which is the same as `<package.Id.Name>_<PublisherId>!App`.

So in theory we can simply combine the strings instead of query the AUMID.