@htcfreek 2022-01-29
> @davidegiacometti 
> The syntax of AMUID is `<package.Id.FamilyName>!App` which is the same as `<package.Id.Name>_<PublisherId>!App`.
> 
> So in theory we can simply combine the strings instead of query the AUMID. I will test this tomorrow.

Damn! The last par of AUMID isn't always "App". This can be set by the dev to an other value. And this value isn't available in `package` element.