@htcfreek 2022-01-29
> > Tested and working as expected!
> > Not sure if there are any side effects on caching AUMID.
> 
> As I know it doesn't change as long as the app exists in the store.

Because it theoretically can change on update (normaly it doesn't), I will add a verification based on the package version.