@htcfreek 2022-01-30
@cinnamon_msft
Do you know if the ApplicationUserModelId for Windows Terminal and Windows Terminal Preview will ever change?