@htcfreek 2022-01-30
@crutkas
If the AUMID only change when the appx package name or developer name/id changes, then I think we can write it fix in the code. This would make the code less complex and the query faster.

(The app name is fix written in code too.)