@htcfreek 2022-01-30
> Nope, the AUMID will not change. Different AUMIDs will be different apps.

Great news. Thank you for the fast answer.